Renamed set1.rs to scancodes.rs. Moved ScancodeSet impls to scancodes.rs
